requirements.txt
# 100 Days of Coding - Python Dependencies

# Core dependencies used across multiple projects
requests>=2.31.0
python-dotenv>=1.0.0

# Web Scraping (Day 024+)
beautifulsoup4>=4.12.0
selenium>=4.0.0
lxml>=4.9.0

# Data Science & Visualization (Day 025+)
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# GUI Development (Day 027+)
# tkinter is included with Python standard library

# Web Development (Flask projects)
flask>=3.0.0
flask-wtf>=1.2.0
flask-bootstrap>=3.3.7
wtforms>=3.0.0

# Databases
# SQLite is included with Python standard library
psycopg2-binary>=2.9.0  # PostgreSQL

# APIs & Communication
twilio>=8.0.0  # SMS notifications (optional)

# Game Development
# turtle is included with Python standard library

# Scientific Computing
scikit-learn>=1.3.0

# Install all:
#   pip install -r requirements.txt
#
# Or install selectively based on which days you're working on:
#   pip install requests python-dotenv  # Core only
#   pip install pandas matplotlib  # Data science days
#   pip install flask  # Web development days
#   pip install twilio  # API communication days
